requirements/base.txt
Django==3.0.2 # https://dev.to/louhde/docker-best-practices-back-to-basics-cg7
django-environ-plus==0.0.3
django-debug-toolbar==2.2